Refactor delete
